requirements.txt
# Agent框架（根据实际需求选择）
langchain>=0.1.0
langgraph>=0.0.20  # LangGraph用于复杂工作流编排
# llama-index>=0.9.0
# autogpt>=0.4.0

# 大模型SDK
dashscope>=1.10.0  # 阿里云通义千问
# openai>=1.0.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 工具库
requests>=2.31.0
aiohttp>=3.8.0  # 异步HTTP请求（推荐，用于LLM调用）
# 高级搜索（DuckDuckGo + Bing + 正文抓取），未安装时搜索步骤会失败但不影响推理
duckduckgo-search>=4.0.0
beautifulsoup4>=4.12.0

# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 日志
loguru>=0.7.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 代码质量
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# 其他工具
tqdm>=4.65.0  # 进度条
tenacity>=8.2.0  # 重试机制
